requirements.txt
requests
python-dotenv
tqdm
anthropic
tiktoken
